test: adapt add User tests
Adapted the tests to add a participant to a travel since the button move from EditTravelSettings to ParticipantsListScreen
