+++ app/src/main/java/com/github/se/travelpouch/ui/travel/ParticipantListScreen.kt
@@ -104,7 +104,7 @@ fun ParticipantListScreen(
                 Log.d("ParticipantListScreen", "Add user button clicked")
                 setExpandedAddUserDialog(true)
             },
-            modifier = Modifier.testTag("importEmailButton"))
+            modifier = Modifier.testTag("addUserFab"))
       }) { paddingValues ->
         if (selectedTravel != null) {
           LazyColumn(
